refactor: extract hot reconfigure and canonicalize profile mode
Move engine ActiveRuntime/reconfigure, the Android session loop, and VPN TUN helpers out of the oversized ControlService, JNI, and VpnService files. Cold and hot Windows VPN attach now share bind_agent_session. Persist mode from frontends, classify Android live mutate once, and keep MASQUE when TUN detach succeeds but system-proxy reapply fails.
